+++ src/config.h
@@ -156,6 +156,7 @@ typedef struct {
         bool enable_perspective_filter;
         GFX_TEXTURE_FILTER texture_filter;
         GFX_TEXTURE_FILTER fbo_filter;
+        bool enable_wireframe;
         bool enable_vsync;
         bool enable_fps_counter;
         float anisotropy_filter;